Protect series sync and add global microphone voice control

+++ app/src/main/java/com/ouropro/player/apps/BaseActivity.java
@@ -73,8 +73,10 @@ public class BaseActivity extends AppCompatActivity {
 
         /* JADX INFO: Access modifiers changed from: private */
         public static /* synthetic */ void lambda$onResponse$0(Response response, Realm realm) {
-            realm.where(SeriesModel.class).findAll().deleteAllFromRealm();
-            realm.insertOrUpdate((Collection<? extends RealmModel>) response.body());
+            Object body = response.body();
+            if (body instanceof Collection && !((Collection) body).isEmpty()) {
+                realm.insertOrUpdate((Collection<? extends RealmModel>) body);
+            }
         }
 
         /* JADX INFO: Access modifiers changed from: private */
@@ -418,8 +420,10 @@ public class BaseActivity extends AppCompatActivity {
 
         /* JADX INFO: Access modifiers changed from: private */
         public static /* synthetic */ void lambda$onResponse$0(Response response, Realm realm) {
-            realm.where(SeriesModel.class).findAll().deleteAllFromRealm();
-            realm.insertOrUpdate((Collection<? extends RealmModel>) response.body());
+            Object body = response.body();
+            if (body instanceof Collection && !((Collection) body).isEmpty()) {
+                realm.insertOrUpdate((Collection<? extends RealmModel>) body);
+            }
         }
 
         /* JADX INFO: Access modifiers changed from: private */
@@ -900,6 +904,12 @@ public class BaseActivity extends AppCompatActivity {
     }
 
     private void getSeriesFromEpisodes(List<EpisodeModel> list) {
+        if (list == null || list.isEmpty()) {
+            if (!this.is_stop) {
+                doNextTask(true);
+            }
+            return;
+        }
         int i;
         List<EpisodeModel> list2;
         List<String> sharedPreferenceSeriesFavNames = this.preferenceHelper.getSharedPreferenceSeriesFavNames();
@@ -1105,8 +1115,10 @@ public class BaseActivity extends AppCompatActivity {
 
     /* JADX INFO: Access modifiers changed from: private */
     public static /* synthetic */ void lambda$getEpisodeModels$15(List list, Realm realm) {
+        if (list == null || list.isEmpty()) {
+            return;
+        }
         realm.where(EpisodeModel.class).findAll().deleteAllFromRealm();
-        realm.where(SeriesModel.class).findAll().deleteAllFromRealm();
         realm.insertOrUpdate(list);
     }
 
@@ -1184,7 +1196,9 @@ public class BaseActivity extends AppCompatActivity {
 
     /* JADX INFO: Access modifiers changed from: private */
     public static /* synthetic */ void lambda$getSeriesFromEpisodes$18(List list, Realm realm) {
-        realm.where(SeriesModel.class).findAll().deleteAllFromRealm();
+        if (list == null || list.isEmpty()) {
+            return;
+        }
         realm.insertOrUpdate(list);
     }
